ai-service/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
psycopg2-binary
